refactor(gc): split run_gc_with_stats diagnostics into gc_stats.rs
gc_runtime.rs keeps the core collector (refcount, three-phase GC,
list/slot management) plus the thin run_gc_with_stats entry point.
run_gc_with_stats_bundle and its diagnostics helpers (edge capture,
object decisions, restoration witnesses) move to gc_stats.rs, a child
module of runtime so it can keep reading the private object table and
GC lists without widening visibility. Public API is unchanged.
